Add dependencies to Morto
You can now have `dependsOnProjects` set on a project, which should be
an array of project names. Say you have projects "A", "B", and
"A-B-integration" which depends on "A" and "B", then it works like
this:
- During setup, if any of "A", "B", or "A-B-integration" changes, then
we'll setup all three, because we know that "A-B-integration" will be
run and it depends on "A" and "B" being set up.
- During other commands, if any of of "A", "B", or "A-B-integration"
changes, then that project will be run, plus also "A-B-integration".

You currently cannot nest dependencies, so you cannot depend on a
project that itself depends on another project.

Will update the README in a separate PR to get this in quickly to
unblock some problems we're seeing in the remix/remix repo.

Test plan: tested this locally on the remix/remix repo.